Refactor all preferences and mapping into one class definition, add email functionality and some cleanup!

+++ cutie.py
@@ -1,9 +1,8 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 # ---------------------------------------------------------------------------
-# created by   : nocturnalbeast
-# created date : 17/12/2021
-# version      : '1.0-multi'
+# author  : nocturnalbeast
+# version : '1.1-multi'
 # ---------------------------------------------------------------------------
 
 """
@@ -32,14 +31,14 @@ import os
 import sys
 import xlsxwriter
 from bs4 import BeautifulSoup
-from rich import print
+from datetime import datetime
 from rich.logging import RichHandler
 from rich.progress import BarColumn, Progress, SpinnerColumn
 from rich.prompt import Confirm
 
-from class_def.alm import ALMConnection
-from class_def.alm_testmapping import ALMTestMapping
-from class_def.preferences import Preferences
+from class_def.alm import ALMConnection, ALMTestMapping
+from class_def.configuration import ConfigStore, write_default_config
+from libs.email import prepare_email, send_email
 from libs.pathops import is_path_exists_or_creatable
 
 
@@ -49,7 +48,7 @@ from libs.pathops import is_path_exists_or_creatable
 QUERY_RESULT_MAX_PER_REQUEST = 100
 
 # this is a user-defined param - tells the script how many connections
-# to make to the server parallelly
+# to make to the server in parallel
 MAX_CONNECTION_THREADS = 5
 
 
@@ -67,12 +66,10 @@ def map_entity_to_test(unprocessed_test_dict, mapping) -> list:
             continue
     # create test record from the mapping
     return [
-        test_field_dict.get(mapping.data[key])
-        if mapping.data[key] != "description"
-        else BeautifulSoup(
-            test_field_dict.get(mapping.data[key], ""), features="lxml"
-        ).text
-        for key in mapping.data.keys()
+        test_field_dict.get(mapping[key])
+        if mapping[key] != "description"
+        else BeautifulSoup(test_field_dict.get(mapping[key], ""), features="lxml").text
+        for key in mapping.keys()
     ]
 
 
@@ -120,17 +117,8 @@ def main():
         "--output",
         action="store",
         type=str,
-        required=True,
         metavar="OUTPUT_FILE",
         help="The Excel file to output to.",
-    )
-    exec_args.add_argument(
-        "-m",
-        "--mapping",
-        action="store",
-        type=str,
-        metavar="MAPPING_FILE",
-        help="The user-defined mapping file that specifies the fields and the relevant column headers (can be YAML/JSON).",
     )
     exec_args.add_argument(
         "-p",
@@ -138,12 +126,24 @@ def main():
         action="store",
         type=str,
         metavar="PREFERENCES_FILE",
-        help="The preferences file that contains the ALM server info (can be YAML/JSON).",
+        help="Path to the preferences file (YAML).",
+    )
+    exec_args.add_argument(
+        "-e",
+        "--email",
+        action="store_true",
+        help="Enables sending an email containing the export once script finishes.",
+    )
+    other_args.add_argument(
+        "-g",
+        "--generate_preferences",
+        action="store_true",
+        help="Generate a default preferences.yaml file to be customized.",
     )
     other_args.add_argument("-h", "--help", action="help", help="Show this cruft.")
     if len(sys.argv) == 1:
         parser.print_help(sys.stderr)
-        sys.exit(1)
+        exit(0)
     args = parser.parse_args()
 
     # setup logging
@@ -160,28 +160,41 @@ def main():
     log = logging.getLogger("cutie-main")
     log.info("Started CUTIE!")
 
+    # check if the generate_preferences flag is set, if yes then call the configuration method to do the same, and exit
+    if args.generate_preferences:
+        write_default_config(os.getcwd() + os.path.sep + "preferences.yaml")
+        exit(0)
+
     # handle output file shennanigans before proceeding
-    output_file_path = os.path.abspath(args.output)
-    if is_path_exists_or_creatable(output_file_path):
-        if os.path.exists(output_file_path):
-            if os.path.isfile(output_file_path):
-                if Confirm.ask(
-                    "Existing file found at the path mentioned. Shall I delete it?"
-                ):
-                    os.unlink(output_file_path)
-                else:
-                    log.error(
-                        "You have cancelled deletion of the file at the path mentioned. Please re-run the script with another path/filename.",
+    if args.output is not None:
+        output_file_path = os.path.abspath(args.output)
+        if is_path_exists_or_creatable(output_file_path):
+            if os.path.exists(output_file_path):
+                if os.path.isfile(output_file_path):
+                    if Confirm.ask(
+                        "Existing file found at the path mentioned. Shall I delete it?"
+                    ):
+                        os.unlink(output_file_path)
+                    else:
+                        log.error(
+                            "You have cancelled deletion of the file at the path mentioned. Please re-run the script with another path/filename.",
+                        )
+                        exit(2)
+                elif os.path.isdir(output_file_path):
+                    output_file_path = (
+                        output_file_path + os.path.sep + "cutie_output.xlsx"
                     )
-                    exit(2)
-            elif os.path.isdir(output_file_path):
-                output_file_path = output_file_path + os.path.sep + "cutie_output.xlsx"
-            else:
-                log.error("Unknown error in path specified!")
-                exit(3)
+                else:
+                    log.error("Unknown error in path specified!")
+                    exit(3)
+        else:
+            log.error("Invalid path and/or insufficient permissions to the path.")
+            exit(4)
     else:
-        log.error("Invalid path and/or insufficient permissions to the path.")
-        exit(4)
+        timestamp = datetime.strftime(datetime.now(), "%Y_%m_%d_%H_%M")
+        output_file_path = os.path.abspath(
+            f"{os.getcwd()}{os.path.sep}export_{timestamp}.xlsx"
+        )
 
     # check preferences file existence, else use fallback/interactive preferences
     if args.preferences is not None:
@@ -190,35 +203,29 @@ def main():
             preferences_file_path
         ):
             log.info("Preferences file found, attempting to retrieve preferences.")
-            pref = Preferences(preferences_file_path)
+            pref = ConfigStore(preferences_file_path)
         else:
-            log.warning("Error in path specified!")
-            pref = Preferences()
+            log.error("Error in path specified!")
+            exit(5)
     else:
-        log.warning("No preferences file specified!")
-        pref = Preferences()
-
-    # check mapping file existence, else use fallback mapping
-    if args.mapping is not None:
-        mapping_file_path = os.path.abspath(args.mapping)
-        if is_path_exists_or_creatable(mapping_file_path) and os.path.isfile(
-            mapping_file_path
-        ):
-            log.info("Mapping file found, attempting to retrieve mapping.")
-            mapping = ALMTestMapping(mapping_file_path)
-        else:
-            log.warning("Error in path specified!")
-            mapping = ALMTestMapping()
-    else:
-        log.warning("No mapping file specified!")
-        mapping = ALMTestMapping()
+        log.error(
+            f'No preferences file specified! Use the command "{os.path.basename(__file__)}" -g to generate a default configuration file!'
+        )
+        exit(6)
 
     # create an authenticated connection using the preferences obtained
     alm_conn = ALMConnection()
-    alm_conn.authenticate(pref)
+    alm_conn.authenticate(
+        pref.alm.webdomain, pref.alm.username, pref.alm.password, pref.alm.https_strict
+    )
+
+    # use the raw data from the preferences to generate the mapping instance
+    test_mapping = ALMTestMapping(
+        pref.raw_data["mapping"].keys(), pref.raw_data["mapping"]
+    )
 
     # setting up some common stuff
-    tests_baseurl = f"{pref.webdomain}/qcbin/rest/domains/{pref.domain}/projects/{pref.project}/tests"
+    tests_baseurl = f"{pref.alm.webdomain}/qcbin/rest/domains/{pref.alm.domain}/projects/{pref.alm.project}/tests"
     req_headers_for_json = {
         "Accept": "application/json",
         "Content-Type": "application/json",
@@ -226,7 +233,7 @@ def main():
 
     # dummy test retrieval to get the number of tests
     res = alm_conn.session.get(
-        tests_baseurl, verify=pref.https_strict, headers=req_headers_for_json
+        tests_baseurl, verify=pref.alm.https_strict, headers=req_headers_for_json
     )
     test_count = json.loads(res.content).get("TotalResults")
     if not isinstance(test_count, int):
@@ -236,7 +243,7 @@ def main():
         log.info(f"Found {test_count} testcases!")
 
     # initialize header for result list
-    all_tests = [mapping.data.keys()]
+    all_tests = [test_mapping.mapping_order]
 
     # start progress bar
     with Progress(
@@ -258,10 +265,10 @@ def main():
                 executor.submit(
                     fetch_test_from_index,
                     alm_conn.session,
-                    mapping,
+                    test_mapping.mapping,
                     tests_baseurl,
                     index,
-                    pref.https_strict,
+                    pref.alm.https_strict,
                 )
                 for index in range(1, test_count, QUERY_RESULT_MAX_PER_REQUEST)
             ]
@@ -272,7 +279,7 @@ def main():
                     task_result = future.result()
                     if task_result is not None:
                         all_tests += task_result
-                except Exception as e:
+                except Exception:
                     log.error("Error trying to get task result!")
                 finally:
                     # advance by max query answers even in case of partial/error response
@@ -285,6 +292,27 @@ def main():
         worksheet = workbook.add_worksheet()
         for idx, data in enumerate(all_tests):
             worksheet.write_row(idx, 0, data)
+
+    # handle email
+    if args.email:
+        from_address, subject, content = prepare_email(pref.email.sender_domain)
+        log.info(
+            f"Sending email has been enabled - sending from address {from_address}..."
+        )
+        status = send_email(
+            from_address,
+            pref.email.to_list,
+            pref.email.cc_list,
+            pref.email.smtp_host,
+            pref.email.smtp_port,
+            subject,
+            content,
+            output_file_path,
+        )
+        if status:
+            log.info("Email sent successfully.")
+        else:
+            log.error("Error encountered in sending email!")
 
 
 if __name__ == "__main__":
